[window] support for DWCAPS_COLOR
when you specifiy DWCAPS_COLOR, the window will have a solid color only.
A surface will not be created, so window->surface is 0.
The color can be specified with Window::SetColor(r,g,b,a).
This is always a "normal" color, NOT premultiplied,
so DSCAPS_PREMULTIPLIED has no effect here.

Currently, this option is only supported by the SaWMan window manager.