Core: Replace FusionCalls by Voodoo Server/Client in core, much better render performance.
This is an experimental branch showing Voodoo performance for the dfbtest_fillrect benchmark.

The secure fusion branch started by adding FusionCalls for object, graphics state, rendering.
FusionCalls are not well optimized for asynchronous calls, streaming, but Voodoo is.

dfbtest_fillrect as a slave via FusionCalls ran at about 50% speed or lower.

In this experimental branch, the same test is running still at 97% speed as a slave, via Voodoo requests!